notification_service/requirements.txt
asgiref==3.5.2
celery==5.2.7
Django==4.0.7
django-computed-property==0.3.0
django-filter==22.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-celery-beat==2.3.0
flower==1.2.0
gunicorn==20.0.4
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv
pytz==2022.2.1
redis==3.4.1
sqlparse==0.4.2
drf-yasg